feat: 鱼桶事件

+++ yangxj96-starter/yangxj96-starter-hikvision/src/main/java/com/yangxj96/saas/starter/hikvision/props/HikvisionProperties.java
@@ -77,7 +77,7 @@ public class HikvisionProperties {
         /**
          * 事件订阅列表
          */
-        private List<Integer> types;
+        private List<Long> types;
 
     }
 }
