+++ yangxj96-starter/yangxj96-starter-hikvision/src/main/java/com/yangxj96/saas/starter/hikvision/request/EventUnSubscriptionDto.java
@@ -19,6 +19,6 @@ public class EventUnSubscriptionDto {
     /**
      * 事件类型
      */
-    private List<Integer> eventTypes;
+    private List<Long> eventTypes;
 
 }
